포스트 상세 조회, 리스트 조회 기능 구현

See merge request hjx66/finalproject_kimheejung_team2!3

+++ src/test/java/com/sns/mutsasns/fixture/TestInfoFixture.java
@@ -6,8 +6,8 @@ import lombok.Setter;
 public class TestInfoFixture {
     public static TestInfo get() {
         TestInfo info = new TestInfo();
-        info.setPostId(1);
-        info.setUserId(1);
+        info.setPostId(1L);
+        info.setUserId(1L);
         info.setUserName("user");
         info.setPassword("password");
         info.setTitle("title");
@@ -18,8 +18,8 @@ public class TestInfoFixture {
     @Getter
     @Setter
     public static class TestInfo {
-        private Integer postId;
-        private Integer userId;
+        private Long postId;
+        private Long userId;
         private String userName;
         private String password;
         private String title;
